# docs/swagger.yaml
definitions:
  v1.ErrorResponse:
    description: Estrutura usada para respostas de erro na API.
    properties:
      error:
        type: string
    type: object
  v1.Person:
    properties:
      age:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
info:
  contact: {}
paths:
  /api/v1/people:
    get:
      description: Obtém uma lista de todas as pessoas do banco de dados MongoDB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.Person'
            type: array
        "500":
          description: 'error: ''Erro ao buscar pessoas'''
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Retorna todas as pessoas
      tags:
      - people
    post:
      consumes:
      - application/json
      description: Insere uma nova pessoa no banco de dados MongoDB
      parameters:
      - description: Pessoa a ser criada
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/v1.Person'
      produces:
      - application/json
      responses:
        "200":
          description: Pessoa criada com sucesso!
          schema:
            $ref: '#/definitions/v1.Person'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Erro ao inserir pessoa no MongoDB
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Cria uma nova pessoa
      tags:
      - people
  /api/v1/products:
    get:
      description: Obtém uma lista de todos os produtos (exemplo, a função está vazia)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Person'
      summary: Retorna todos os produtos
      tags:
      - products
    post:
      description: Cria um produto (exemplo, a função está vazia)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Person'
      summary: Cria um novo produto
      tags:
      - products
swagger: "2.0"